possibility to extend tests in client projects
